improved Linux makefile compilation, renamed collision.h to collisions.h for better cross-platform support (meow aap), compilation fixes pack meow

+++ tools/IIItest/src/World.h
@@ -30,5 +30,5 @@ public:
 	static CPtrList &GetMovingEntityList(void) { return ms_listMovingEntityPtrs; }
 	static void ClearScanCodes(void);
 
-	static int GetSectorIndex(float f) { return (f + 2000.0f)/40.0f; }
+	static int GetSectorIndex(float f) { return (int)( (f + 2000.0f)/40.0f ); }
 };
